refactor(relay): extract the TURN state machine into relay-proto
The fuzz harness simulates a relay and therefore depended on the whole
`firezone-relay` crate, dragging the portal channel, telemetry, sockets
and the eBPF toolchain into every fuzz build for the sake of a sans-IO
state machine. Split that state machine, its auth logic and its metric
definitions into `relay-proto`, mirroring `tunnel-proto`.

`firezone-relay` re-exports the moved items, so the binary is unchanged.
The `ebpf` feature existed so library consumers could opt out of
building the XDP program; the only such consumer now depends on
`relay-proto` instead, so the feature is gone and `aya` is an ordinary
dependency of the binary crate.
