+++ rust/relay/proto/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "relay-proto"
+version = "0.1.0"
+edition = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+base64 = { workspace = true, features = ["std"] }
+bimap = { workspace = true }
+bytecodec = { workspace = true }
+derive_more = { workspace = true, features = ["from"] }
+hex = { workspace = true }
+logging = { workspace = true }
+once_cell = { workspace = true }
+opentelemetry = { workspace = true, features = ["metrics"] }
+proptest = { workspace = true, optional = true }
+rand = { workspace = true }
+secrecy = { workspace = true }
+sha2 = { workspace = true }
+smallvec = { workspace = true }
+stun_codec = { workspace = true }
+thiserror = { workspace = true }
+tracing = { workspace = true }
+uuid = { workspace = true, features = ["v4"] }
+
+[dev-dependencies]
+difference = { workspace = true }
+test-strategy = { workspace = true }
+
+[[test]]
+name = "regression"
+path = "tests/regression.rs"
+required-features = ["proptest"]
+
+[lints]
+workspace = true
