+++ rust/relay/server/build.rs
@@ -3,14 +3,6 @@ fn main() -> anyhow::Result<()> {
     use anyhow::Context as _;
     use aya_build::cargo_metadata::{MetadataCommand, Package};
 
-    // Build scripts don't receive feature cfgs, only the `CARGO_FEATURE_<name>`
-    // env var for each activated feature. When the `ebpf` feature is disabled we
-    // skip compiling the eBPF program entirely, which avoids the dependency on the
-    // pinned nightly toolchain and `bpf-linker`.
-    if std::env::var_os("CARGO_FEATURE_EBPF").is_none() {
-        return Ok(());
-    }
-
     let package = MetadataCommand::new()
         .no_deps()
         .exec()
