+++ rust/tests/fuzz/Cargo.toml
@@ -31,7 +31,6 @@ connlib-model = { workspace = true }
 derive_more = { workspace = true, features = ["debug"] }
 dns-over-tcp = { workspace = true }
 dns-types = { workspace = true }
-firezone-relay = { workspace = true }
 flow-tracker = { workspace = true }
 hex = { workspace = true }
 ip-packet = { workspace = true, features = ["arbitrary"] }
@@ -41,6 +40,7 @@ itertools = { workspace = true, features = ["use_std"] }
 l3-tcp = { workspace = true }
 libfuzzer-sys = { workspace = true }
 rand = { workspace = true }
+relay-proto = { workspace = true }
 secrecy = { workspace = true, features = ["serde"] }
 serde_json = { workspace = true }
 sha2 = { workspace = true }
